fixes update-django-db
Prevents script to execute 'git reset --hard' command when there are unstaged changes.

+++ bin/update-django-db.sh
@@ -2,7 +2,7 @@
 
 # Path:        ~/git/lab/lex-crawler/update-db.sh
 # Created:     2019-11-06, 22:04    @lenovo
-# Last update: 2019-11-09, 20:50:31 @lenovo
+# Last update: 2019-11-11, 13:30:19 @x200
 
 DB=db.sqlite3
 
@@ -11,15 +11,17 @@ NORMAL=$(tput sgr0)
 GRE=$(tput setaf 10)
 RED=$(tput setaf 1)
 
-git_reset=false
+force_git=false
+reset_git=false
 superuser=false
+
 
 print_green() { echo "${BOLD}${GRE}>>> ${1} ${NORMAL}"; }
 print_red() { echo "${BOLD}${RED}>>> ${1} ${NORMAL}"; }
 
 usage(){
     cat <<EOF
-Usage: update-db.sh [-hbdrg]
+Usage: update-db.sh [-hbdrfg]
 
 Automate database cleanup in Django. Execute in venv.
 
@@ -28,7 +30,10 @@ Options:
   -b   backup current database in bak/ directory (excludes -d)
   -d   delete current database
   -r   remove migrations
-  -g   execute "git reset --hard" (use with -r)
+  -f   force git reset even when there are uncommited changes
+       (works only with -g flag)
+  -g   execute "git reset --hard" (use with -r) if there are not
+       uncommitted changes (unless flag -f provided)
 
 EOF
     exit 0
@@ -55,7 +60,7 @@ check_venv(){
 }
 
 parse_args(){
-    while getopts "hbdrge" option; do
+    while getopts "hbdfrge" option; do
         case "${option}" in
             h) usage; exit 0 ;;
             b) print_green "Making database backup"
@@ -63,10 +68,11 @@ parse_args(){
                [[ -f $DB ]] && mv $DB bak/$DB$(date +"_%d%m%Y_%H%M")
                ;;
             d) [[ -f $DB ]] && print_red "Deleting current database"; rm $DB ;;
+            f) force_git=true ;;
             r) print_red "Deleting migrations"
                rm */migrations/000*
                ;;
-            g) git_reset=true ;;
+            g) reset_git=true ;;
             e) exit 0 ;;
         esac
     done
@@ -86,8 +92,18 @@ create_superuser(){
     fi
 }
 
-gitreset(){
-    if [ $git_reset = true ]; then
+check_git(){
+    status=$(git status | grep "Changes not staged")
+    if [[ $(git status | grep "Changes not staged") ]] && [[ $force_git != "true" ]]; then
+        print_red "you have unstaged changes, stage them or provide flag -f"
+        exit 1
+    fi
+}
+
+reset_git(){
+    check_git
+
+    if [ $reset_git = true ]; then
         print_red "Resetting git to last commit"
         git reset --hard
         migrations "Re-making migrations" "Re-migrating"
@@ -100,7 +116,7 @@ main(){
     ! [[ -f $DB ]] && superuser=true
     migrations "Making migrations" "Migrating"
     create_superuser
-    gitreset
+    reset_git
     print_green "Done!"
 }
